+++ src/app/chatbot-design-studio/services/intent.service.ts
@@ -3,7 +3,7 @@ import { Subject, BehaviorSubject } from 'rxjs';
 import { LoggerService } from 'src/chat21-core/providers/abstract/logger.service';
 import { v4 as uuidv4 } from 'uuid';
 
-import { ActionReply, ActionAgent, ActionAssignFunction, ActionAssignVariable, ActionChangeDepartment, ActionClose, ActionDeleteVariable, ActionEmail, ActionHideMessage, ActionIntentConnected, ActionJsonCondition, ActionOnlineAgent, ActionOpenHours, ActionRandomReply, ActionReplaceBot, ActionWait, ActionWebRequest, Command, Wait, Message, Expression, Action, ActionAskGPT, ActionWhatsappAttribute, ActionWhatsappStatic, ActionWebRequestV2, ActionGPTTask, ActionCaptureUserReply, ActionQapla, ActionCondition, ActionMake, ActionAssignVariableV2, ActionHubspot, ActionCode, ActionReplaceBotV2, ActionAskGPTV2, ActionCustomerio, ActionVoice, ActionBrevo, Attributes, ActionN8n, ActionGPTAssistant } from 'src/app/models/action-model';
+import { ActionReply, ActionAgent, ActionAssignFunction, ActionAssignVariable, ActionChangeDepartment, ActionClose, ActionDeleteVariable, ActionEmail, ActionHideMessage, ActionIntentConnected, ActionJsonCondition, ActionOnlineAgent, ActionOpenHours, ActionRandomReply, ActionReplaceBot, ActionWait, ActionWebRequest, Command, Wait, Message, Expression, Action, ActionAskGPT, ActionWhatsappAttribute, ActionWhatsappStatic, ActionWebRequestV2, ActionGPTTask, ActionCaptureUserReply, ActionQapla, ActionCondition, ActionMake, ActionAssignVariableV2, ActionHubspot, ActionCode, ActionReplaceBotV2, ActionAskGPTV2, ActionCustomerio, ActionVoice, ActionBrevo, Attributes, ActionN8n, ActionGPTAssistant, ActionReplyV2 } from 'src/app/models/action-model';
 import { Intent } from 'src/app/models/intent-model';
 import { FaqService } from 'src/app/services/faq.service';
 import { FaqKbService } from 'src/app/services/faq-kb.service';
@@ -794,6 +794,15 @@ export class IntentService {
       command.message = new Message('text', 'A chat message will be sent to the visitor');
       action.attributes.commands.push(command);
     }
+    if(typeAction === TYPE_ACTION.REPLYV2){
+      action = new ActionReplyV2();
+      action.noInputTimeout = 10000;
+      let commandWait = new Wait();
+      action.attributes.commands.push(commandWait);
+      let command = new Command(TYPE_COMMAND.MESSAGE);
+      command.message = new Message('text', 'A chat message will be sent to the visitor');
+      action.attributes.commands.push(command);
+    }
     if(typeAction === TYPE_ACTION.RANDOM_REPLY){
       action = new ActionRandomReply();
       let commandWait = new Wait();
@@ -873,13 +882,13 @@ export class IntentService {
     }
     if(typeAction === TYPE_ACTION.ASKGPT){
       action = new ActionAskGPT();
-      action.question = '{{last_user_text}}'
+      action.question = '{{lastUserText}}'
       action.assignReplyTo = 'kb_reply';
       action.assignSourceTo = 'kb_source';
     }
     if(typeAction === TYPE_ACTION.ASKGPTV2) {
       action = new ActionAskGPTV2();
-      action.question = '{{last_user_text}}'
+      action.question = '{{lastUserText}}'
       action.assignReplyTo = 'kb_reply';
       action.assignSourceTo = 'kb_source';
       action.max_tokens = 512;
@@ -898,7 +907,8 @@ export class IntentService {
     }
     if(typeAction === TYPE_ACTION.GPT_ASSISTANT){
       action = new ActionGPTAssistant();
-      action.threadIdAttribute = 'firstThread'
+      action.prompt ='{{lastUserText}}';
+      action.threadIdAttribute = 'firstThread';
       action.assignResultTo = 'assistantReply';
       action.assignErrorTo = 'assistantError';
     }