+++ CHANGELOG.md
@@ -6,6 +6,13 @@
     *Giovanni Troisi*  
 ### **Copyrigth**: *Tiledesk SRL*
 
+### 1.8.0-rc.2
+👉 **changed**: set default values for cds-action-gpt-assistant and cds-action-reply-v2
+
+### 1.8.0-rc.1
+👉 **added**: cds-action-reply-v2
+👉 **changed**: 'add action' button into cds-intent preview component UI changed
+
 ### 1.7.0 in PROD
 
 ### 1.7.0-rc.1